Consolidate Slack parameter (#548)
